rename HandlePostWebsites to addWebsiteHandler
